+++ docs.md
@@ -1,0 +1,76 @@
+# Related Topics
+
+
+## 1. Alcohol Dataset
+
+### Dataset Information
+This dataset contains information about alcohol consumption per capita across different countries. The data is sourced from the OECD and can be accessed on Kaggle: [OECD Alcohol Consumption Per Capita](https://www.kaggle.com/datasets/thedevastator/oecd-alcohol-consumption-per-capita/data).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+## 2. Coffee Dataset
+
+### Dataset Information
+This dataset is a valuable asset for those interested in exploring how coffee consumption varies across different regions and over time.  [Worldwide Coffee Habits Dataset](https://www.kaggle.com/datasets/waqi786/worldwide-coffee-habits-dataset/data).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+### Note
+I couldn't find the mapping of country codes and countries. But I think it can be inetresting task to guess the countries based on the coffee + happiness data. What do you think?
+
+## 3. Divorce Rates Data: Should You Get Married?
+
+### Dataset Information
+In this dataset, we can spot facts about divorces, and trends across countries - continents.
+Mainly we have the marriage, divorce rates and the percentage score as well as numbers of divorces (given by the source and year as provided): [Divorce statistics](https://www.kaggle.com/datasets/johnny1994/divorce-rates-data-should-you-get-married).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+## 4. Meat Consumption Dataset
+
+### Dataset Information
+This dataset provides a rich resource for understanding the historical and cultural context of meat consumption around the world and its impact on human health, the environment, and food systems. The dataset can be accessed on Kaggle: [Meat Consumption Dataset](https://www.kaggle.com/datasets/ulrikthygepedersen/meat-consumption).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+## 5. Years of Schooling Worldwide
+
+### Dataset Information
+This dataset presents a comprehensive historical record of Expected Years of Schooling data on a global scale. It encompasses vital columns such as ISO3 (the ISO3 code for each country/territory), Country (the name of the country or territory), Continent (the continent where the country is located), Hemisphere (the hemisphere in which the country is situated), Human Development Groups, UNDP Developing Regions, HDI Rank (2021) denoting the Human Development Index Rank for the year 2021 and Expected Years of Schooling values spanning from 1990 to 2021. This dataset provides valuable insights into the evolution and disparities in educational prospects across the world.
+
+: [Schooling Datset](https://www.kaggle.com/datasets/iamsouravbanerjee/years-of-schooling-worldwide).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+## 6. StarBucks
+
+### Dataset Information
+This dataset includes a record for every Starbucks or subsidiary store location currently in operation as of February 2017 [Starbucks 2017](https://www.kaggle.com/datasets/starbucks/store-locations).
+This dataset includes a record for every Starbucks or subsidiary store location currently in operation as of Nov 2021 [Starbucks 2021](https://www.kaggle.com/datasets/kukuroo3/starbucks-locations-worldwide-2021-version).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+  
+## 7. Weather
+
+### Dataset Information
+This dataset offers a comprehensive collection of Daily weather readings from major cities around the world. In the first release, it included only capitals, but now it also adds main cities worldwide and hourly data as well, making up to ~1250 cities. Some locations provide historical data tracing back to January 2, 1833, giving users a deep dive into long-term weather patterns and their evolution.: [Weather Datset](https://www.kaggle.com/datasets/guillemservera/global-daily-climate-data?select=daily_weather.parquet).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
+
+### NOTE
+Mko: Please review this data before adding it. I think there can be some inteereting connections. But the data file is very big.
+
+## 8. Worldwide Average IQ Levels
+
+### Dataset Information
+The dataset file contains a comprehensive dataset that explores the relationship between average IQ levels and various socioeconomic factors across different countries: [IQ Datset](https://www.kaggle.com/datasets/abhijitdahatonde/worldwide-average-iq-levels/data).
+
+### Interesting Correlation
+If you find any interesting correlations or insights while analyzing this dataset, please document them here.
