--- related_topics/alcohol_consumption/docs_alcohol.md
@@ -1,7 +1,0 @@
-# Alcohol Dataset
-
-## Dataset Information
-This dataset contains information about alcohol consumption per capita across different countries. The data is sourced from the OECD and can be accessed on Kaggle: [OECD Alcohol Consumption Per Capita](https://www.kaggle.com/datasets/thedevastator/oecd-alcohol-consumption-per-capita/data).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
