--- related_topics/coffee_consumption/docs_coffee.md
@@ -1,10 +1,0 @@
-# Alcohol Dataset
-
-## Dataset Information
-This dataset is a valuable asset for those interested in exploring how coffee consumption varies across different regions and over time.  [Worldwide Coffee Habits Dataset](https://www.kaggle.com/datasets/waqi786/worldwide-coffee-habits-dataset/data).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
-
-## Note
-I couldn't find the mapping of country codes and countries. But I think it can be inetresting task to guess the countries based on the coffee + happiness data. What do you think?
