--- related_topics/docs.md
@@ -1,23 +1,0 @@
-# Related Topics
-
-1. **Alcohol Consumption**
-   - **Directory:** [related_topics/alcohol_consumption](/related_topics/alcohol_consumption/docs_alcohol.md)
-
-2. **Coffee Consumption**
-   - **Directory:** [related_topics/coffee_consumption](/related_topics/coffee_consumption/docs_coffee.md)
-
-3. **Meat Consumption**
-   - **Directory:** [related_topics/meat_consumption](/related_topics/meat_consumption/docs_meat.md)
-
-4. **Worldwide Average IQ Levels**
-   - **Directory:** [related_topics/wordwide_iq](/related_topics/wordwide_iq/docs_iq.md) (not time series)
-
-5. **Years of Schooling Worldwide**
-   - **Directory:** [related_topics/schooling](/related_topics/schooling/docs_schooling.md)
-
-6. **Weather**
-   - **Directory:** [related_topics/schooling](/related_topics/weather/docs_weather.md)
-  
-8. **Divorce By Country**
-   - **Directory:** [related_topics/marriage_divorce](/related_topics/marriage_divorce/docs_marriage.md) (not time series)
-
