--- related_topics/marriage_divorce/docs_marriage.md
@@ -1,8 +1,0 @@
-# Divorce Rates Data: Should You Get Married?
-
-## Dataset Information
-In this dataset, we can spot facts about divorces, and trends across countries - continents.
-Mainly we have the marriage, divorce rates and the percentage score as well as numbers of divorces (given by the source and year as provided): [Divorce statistics](https://www.kaggle.com/datasets/johnny1994/divorce-rates-data-should-you-get-married).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
