--- related_topics/meat_consumption/docs_meat.md
@@ -1,7 +1,0 @@
-# Meat Consumption Dataset
-
-## Dataset Information
-This dataset provides a rich resource for understanding the historical and cultural context of meat consumption around the world and its impact on human health, the environment, and food systems. The dataset can be accessed on Kaggle: [Meat Consumption Dataset](https://www.kaggle.com/datasets/ulrikthygepedersen/meat-consumption).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
