--- related_topics/schooling/docs_schoolng.md
@@ -1,9 +1,0 @@
-# Years of Schooling Worldwide
-
-## Dataset Information
-This dataset presents a comprehensive historical record of Expected Years of Schooling data on a global scale. It encompasses vital columns such as ISO3 (the ISO3 code for each country/territory), Country (the name of the country or territory), Continent (the continent where the country is located), Hemisphere (the hemisphere in which the country is situated), Human Development Groups, UNDP Developing Regions, HDI Rank (2021) denoting the Human Development Index Rank for the year 2021 and Expected Years of Schooling values spanning from 1990 to 2021. This dataset provides valuable insights into the evolution and disparities in educational prospects across the world.
-
-: [Schooling Datset](https://www.kaggle.com/datasets/iamsouravbanerjee/years-of-schooling-worldwide).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
