--- related_topics/starbucks/docs_starbucks.md
@@ -1,8 +1,0 @@
-# Worldwide Average IQ Levels
-
-## Dataset Information
-This dataset includes a record for every Starbucks or subsidiary store location currently in operation as of February 2017 [Starbucks 2017](https://www.kaggle.com/datasets/starbucks/store-locations).
-This dataset includes a record for every Starbucks or subsidiary store location currently in operation as of Nov 2021 [Starbucks 2021](https://www.kaggle.com/datasets/kukuroo3/starbucks-locations-worldwide-2021-version).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
