--- related_topics/weather/docs_weather.md
@@ -1,10 +1,0 @@
-# Weather
-
-## Dataset Information
-This dataset offers a comprehensive collection of Daily weather readings from major cities around the world. In the first release, it included only capitals, but now it also adds main cities worldwide and hourly data as well, making up to ~1250 cities. Some locations provide historical data tracing back to January 2, 1833, giving users a deep dive into long-term weather patterns and their evolution.: [Weather Datset](https://www.kaggle.com/datasets/guillemservera/global-daily-climate-data?select=daily_weather.parquet).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
-
-## NOTE
-Mko: Please review this data before adding it. I think there can be some inteereting connections. But the data file is very big.
