--- related_topics/wordwide_iq/docs_iq.md
@@ -1,7 +1,0 @@
-# Worldwide Average IQ Levels
-
-## Dataset Information
-The dataset file contains a comprehensive dataset that explores the relationship between average IQ levels and various socioeconomic factors across different countries: [IQ Datset](https://www.kaggle.com/datasets/abhijitdahatonde/worldwide-average-iq-levels/data).
-
-## Interesting Correlation
-If you find any interesting correlations or insights while analyzing this dataset, please document them here.
